+++ baiducloud-python-sdk-iam/ChangeLog.md
@@ -1,2 +1,8 @@
+2026-07-21 Version: 0.0.2
+- 涉及产品: IAM，创建永久API Key响应新增id、tokenId、services等字段
+- 涉及产品: IAM，更新/查询/解码永久API Key响应新增访问控制及时间字段
+- 涉及产品: IAM，获取永久API Key列表及删除接口新增状态与分页字段
+- 涉及产品: IAM，创建短期API Key响应新增userId、token及有效期字段
+
 2026-06-30 Version: 0.0.1
 - 涉及产品: IAM，SDK版本更新